rename eventlog into windowsevents for clarity
