+++ pkg/logs/agent.go
@@ -9,10 +9,10 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/logs/auditor"
 	"github.com/DataDog/datadog-agent/pkg/logs/config"
 	"github.com/DataDog/datadog-agent/pkg/logs/input/container"
-	"github.com/DataDog/datadog-agent/pkg/logs/input/eventlog"
 	"github.com/DataDog/datadog-agent/pkg/logs/input/journald"
 	"github.com/DataDog/datadog-agent/pkg/logs/input/listener"
 	"github.com/DataDog/datadog-agent/pkg/logs/input/tailer"
+	"github.com/DataDog/datadog-agent/pkg/logs/input/windowsevent"
 	"github.com/DataDog/datadog-agent/pkg/logs/message"
 	"github.com/DataDog/datadog-agent/pkg/logs/pipeline"
 	"github.com/DataDog/datadog-agent/pkg/logs/restart"
@@ -27,13 +27,13 @@ import (
 // |                                                        |
 // + ------------------------------------------------------ +
 type Agent struct {
-	auditor           *auditor.Auditor
-	containersScanner *container.Scanner
-	eventlogLauncher  *eventlog.Launcher
-	filesScanner      *tailer.Scanner
-	networkListener   *listener.Listener
-	journaldLauncher  *journald.Launcher
-	pipelineProvider  pipeline.Provider
+	auditor              *auditor.Auditor
+	containersScanner    *container.Scanner
+	windowsEventLauncher *windowsevent.Launcher
+	filesScanner         *tailer.Scanner
+	networkListener      *listener.Listener
+	journaldLauncher     *journald.Launcher
+	pipelineProvider     pipeline.Provider
 }
 
 // NewAgent returns a new Agent
@@ -56,16 +56,16 @@ func NewAgent(sources *config.LogSources) *Agent {
 	networkListeners := listener.New(validSources, pipelineProvider)
 	filesScanner := tailer.New(validSources, config.LogsAgent.GetInt("logs_config.open_files_limit"), pipelineProvider, auditor, tailer.DefaultSleepDuration)
 	journaldLauncher := journald.New(validSources, pipelineProvider, auditor)
-	eventlogLauncher := eventlog.New(validSources, pipelineProvider, auditor)
+	windowsEventLauncher := windowsevent.New(validSources, pipelineProvider, auditor)
 
 	return &Agent{
-		auditor:           auditor,
-		containersScanner: containersScanner,
-		eventlogLauncher:  eventlogLauncher,
-		filesScanner:      filesScanner,
-		journaldLauncher:  journaldLauncher,
-		networkListener:   networkListeners,
-		pipelineProvider:  pipelineProvider,
+		auditor:              auditor,
+		containersScanner:    containersScanner,
+		windowsEventLauncher: windowsEventLauncher,
+		filesScanner:         filesScanner,
+		journaldLauncher:     journaldLauncher,
+		networkListener:      networkListeners,
+		pipelineProvider:     pipelineProvider,
 	}
 }
 
@@ -79,7 +79,7 @@ func (a *Agent) Start() {
 		a.networkListener,
 		a.containersScanner,
 		a.journaldLauncher,
-		a.eventlogLauncher,
+		a.windowsEventLauncher,
 	)
 }
 
@@ -92,7 +92,7 @@ func (a *Agent) Stop() {
 			a.networkListener,
 			a.containersScanner,
 			a.journaldLauncher,
-			a.eventlogLauncher,
+			a.windowsEventLauncher,
 		),
 		a.pipelineProvider,
 		a.auditor,
