+++ pkg/logs/input/windowsevent/eventlog.c
